Consolidate telemetry tests into telemetry.spec.ts and cover getLogger
test/internal/internalAPI.spec.ts contained only telemetry tests and
nothing from internalAPIs.ts. PR #3123 added internalAPIs.spec.ts, so
the repo had two near-identically-named files covering entirely
different modules. Rename the misnamed file to telemetry.spec.ts to
match the module under test.

Also add coverage for getLogger, which had none: assert the returned
logger is namespaced under the top-level teamsJs namespace and that
nesting via .extend() composes.
